refactor(board-exception): use base-package as `nettee.board.exception`
