+++ services/board/api/exception/src/main/java/nettee/board/exception/BoardCommandException.java
@@ -1,4 +1,4 @@
-package nettee.board;
+package nettee.board.exception;
 
 import nettee.common.CustomException;
 import nettee.common.ErrorCode;
